Unify CRD functions for CNP and ANP
