+++ hack/generate-manifest.sh
@@ -28,8 +28,8 @@ Generate a YAML manifest for Antrea using Kustomize and print it to stdout.
         --cloud                       Generate a manifest appropriate for running Antrea in Public Cloud
         --ipsec                       Generate a manifest with IPSec encryption of tunnel traffic enabled
         --proxy                       Generate a manifest with Antrea proxy enabled
-        --np                          Generate a manifest with ClusterNetworkPolicy related CRDs enabled
-        --anp                         Generate a manifest with Namespaced Antrea NetworkPolicy CRDs enabled
+        --np                          Generate a manifest with ClusterNetworkPolicy and AntreaNetworkPolicy feature gate enabled
+        --core-crd                    Generate a manifest with ExternalEntity CRD enabled
         --keep                        Debug flag which will preserve the generated kustomization.yml
         --tun (geneve|vxlan|gre|stt)  Choose encap tunnel type from geneve, gre, stt and vxlan (default is geneve)
         --help, -h                    Print this message and exit
@@ -54,7 +54,7 @@ KIND=false
 IPSEC=false
 PROXY=false
 NP=false
-ANP=false
+CORE_CRD=false
 KEEP=false
 ENCAP_MODE=""
 CLOUD=""
@@ -93,8 +93,8 @@ case $key in
     NP=true
     shift
     ;;
-    --anp)
-    ANP=true
+    --core-crd)
+    CORE_CRD=true
     shift
     ;;
     --keep)
@@ -188,10 +188,8 @@ fi
 if $NP; then
     sed -i.bak -E "s/^[[:space:]]*#[[:space:]]*ClusterNetworkPolicy[[:space:]]*:[[:space:]]*[a-z]+[[:space:]]*$/  ClusterNetworkPolicy: true/" antrea-controller.conf
     sed -i.bak -E "s/^[[:space:]]*#[[:space:]]*ClusterNetworkPolicy[[:space:]]*:[[:space:]]*[a-z]+[[:space:]]*$/  ClusterNetworkPolicy: true/" antrea-agent.conf
-fi
-
-if $ANP; then
     sed -i.bak -E "s/^[[:space:]]*#[[:space:]]*AntreaNetworkPolicy[[:space:]]*:[[:space:]]*[a-z]+[[:space:]]*$/  AntreaNetworkPolicy: true/" antrea-controller.conf
+    sed -i.bak -E "s/^[[:space:]]*#[[:space:]]*AntreaNetworkPolicy[[:space:]]*:[[:space:]]*[a-z]+[[:space:]]*$/  AntreaNetworkPolicy: true/" antrea-agent.conf
 fi
 
 if [[ $ENCAP_MODE != "" ]]; then
@@ -226,17 +224,14 @@ if $IPSEC; then
     cd ..
 fi
 
-if $NP; then
+if $CORE_CRD; then
     mkdir np && cd np
-    cp ../../patches/np/*.yml .
-    cp ../../base/security-crds.yml .
+    cp ../../patches/np/npRbac.yml .
     cp ../../base/core-crds.yml .
     touch kustomization.yml
     $KUSTOMIZE edit add base $BASE
-    # add RBAC to antrea-controller for ANP and CNP CRD access.
+    # add RBAC to antrea-controller for ExternalEntity CRD access.
     $KUSTOMIZE edit add patch npRbac.yml
-    # create NetworkPolicy related CRDs.
-    $KUSTOMIZE edit add resource security-crds.yml
     # create ExternalEntity related CRDs.
     $KUSTOMIZE edit add resource core-crds.yml
     BASE=../np
