+++ pkg/controller/networkpolicy/antreanetworkpolicy.go
@@ -15,22 +15,12 @@
 package networkpolicy
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/klog"
 
 	"github.com/vmware-tanzu/antrea/pkg/apis/networking"
 	secv1alpha1 "github.com/vmware-tanzu/antrea/pkg/apis/security/v1alpha1"
 	antreatypes "github.com/vmware-tanzu/antrea/pkg/controller/types"
-)
-
-var (
-	// matchAllPodsPeerCrd is a secv1alpha1.NetworkPolicyPeer matching all
-	// Pods from all Namespaces.
-	matchAllPodsPeerCrd = secv1alpha1.NetworkPolicyPeer{
-		NamespaceSelector: &metav1.LabelSelector{},
-	}
 )
 
 // addANP receives Antrea NetworkPolicy ADD events and creates resources
@@ -53,7 +43,7 @@ func (n *NetworkPolicyController) addANP(obj interface{}) {
 func (n *NetworkPolicyController) updateANP(old, cur interface{}) {
 	defer n.heartbeat("updateANP")
 	curNP := cur.(*secv1alpha1.NetworkPolicy)
-	klog.Infof("Processing Antrea NetworkPolicy %s/%s UPDATE event", curCNP.Namespace, curCNP.Name)
+	klog.Infof("Processing Antrea NetworkPolicy %s/%s UPDATE event", curNP.Namespace, curNP.Name)
 	// Update an internal NetworkPolicy, corresponding to this NetworkPolicy and
 	// enqueue task to internal NetworkPolicy Workqueue.
 	curInternalNP := n.processAntreaNetworkPolicy(curNP)
@@ -126,97 +116,6 @@ func (n *NetworkPolicyController) deleteANP(old interface{}) {
 	n.deleteDereferencedAddressGroups(oldInternalNP)
 }
 
-// toAntreaServicesForCRD converts a secv1alpha1.NetworkPolicyPort object to an
-// Antrea Service object.
-func toAntreaServicesForCRD(npPorts []secv1alpha1.NetworkPolicyPort) []networking.Service {
-	var antreaServices []networking.Service
-	for _, npPort := range npPorts {
-		antreaService := networking.Service{
-			Protocol: toAntreaProtocol(npPort.Protocol),
-			Port:     npPort.Port,
-		}
-		antreaServices = append(antreaServices, antreaService)
-	}
-	return antreaServices
-}
-
-// toAntreaIPBlockForCRD converts a secv1alpha1.IPBlock to an Antrea IPBlock.
-func toAntreaIPBlockForCRD(ipBlock *secv1alpha1.IPBlock) (*networking.IPBlock, error) {
-	// Convert the allowed IPBlock to networkpolicy.IPNet.
-	ipNet, err := cidrStrToIPNet(ipBlock.CIDR)
-	if err != nil {
-		return nil, err
-	}
-	antreaIPBlock := &networking.IPBlock{
-		CIDR: *ipNet,
-		// secv1alpha.IPBlock does not have the Except slices.
-		Except: []networking.IPNet{},
-	}
-	return antreaIPBlock, nil
-}
-
-func (n *NetworkPolicyController) toAntreaPeerForCRD(peers []secv1alpha1.NetworkPolicyPeer, np *secv1alpha1.NetworkPolicy, dir networking.Direction) *networking.NetworkPolicyPeer {
-	var addressGroups []string
-	// Empty NetworkPolicyPeer is supposed to match all addresses.
-	// It's treated as an IPBlock "0.0.0.0/0".
-	if len(peers) == 0 {
-		// For an ingress Peer, skip adding the AddressGroup matching all Pods
-		// because in case of ingress Rule, the named Port resolution happens on
-		// Pods in AppliedToGroup.
-		if dir == networking.DirectionIn {
-			return &matchAllPeer
-		}
-		// For an egress Peer, create an AddressGroup matching all Pods in all
-		// Namespaces such that it can be used to resolve named Ports. This
-		// AddressGroup is set in the NetworkPolicyPeer of matchAllPeer.
-		allPodsGroupUID := n.createAddressGroupForCRD(matchAllPodsPeerCrd, np)
-		podsPeer := matchAllPeer
-		addressGroups = append(addressGroups, allPodsGroupUID)
-		podsPeer.AddressGroups = addressGroups
-		return &podsPeer
-	}
-	var ipBlocks []networking.IPBlock
-	for _, peer := range peers {
-		// A secv1alpha1.NetworkPolicyPeer will either have an IPBlock or a
-		// podSelector and/or namespaceSelector set.
-		if peer.IPBlock != nil {
-			ipBlock, err := toAntreaIPBlockForCRD(peer.IPBlock)
-			if err != nil {
-				klog.Errorf("Failure processing Antrea NetworkPolicy %s/%s IPBlock %v: %v", np.Namespace, np.Name, peer.IPBlock, err)
-				continue
-			}
-			ipBlocks = append(ipBlocks, *ipBlock)
-		} else {
-			normalizedUID := n.createAddressGroupForCRD(peer, np)
-			addressGroups = append(addressGroups, normalizedUID)
-		}
-	}
-	return &networking.NetworkPolicyPeer{AddressGroups: addressGroups, IPBlocks: ipBlocks}
-}
-
-// createAddressGroupForCRD creates an AddressGroup object corresponding to a
-// secv1alpha1.NetworkPolicyPeer object in Antrea NetworkPolicyRule. This
-// function simply creates the object without actually populating the
-// PodAddresses as the affected Pods are calculated during sync process.
-func (n *NetworkPolicyController) createAddressGroupForCRD(peer secv1alpha1.NetworkPolicyPeer, np *secv1alpha1.NetworkPolicy) string {
-	groupSelector := toGroupSelector(np.Namespace, peer.PodSelector, peer.NamespaceSelector)
-	normalizedUID := getNormalizedUID(groupSelector.NormalizedName)
-	// Get or create an AddressGroup for the generated UID.
-	_, found, _ := n.addressGroupStore.Get(normalizedUID)
-	if found {
-		return normalizedUID
-	}
-	// Create an AddressGroup object per Peer object.
-	addressGroup := &antreatypes.AddressGroup{
-		UID:      types.UID(normalizedUID),
-		Name:     normalizedUID,
-		Selector: *groupSelector,
-	}
-	klog.V(2).Infof("Creating new AddressGroup %s with selector (%s)", addressGroup.Name, addressGroup.Selector.NormalizedName)
-	n.addressGroupStore.Create(addressGroup)
-	return normalizedUID
-}
-
 // processAntreaNetworkPolicy creates an internal NetworkPolicy instance
 // corresponding to the secv1alpha1.NetworkPolicy object. This method
 // does not commit the internal NetworkPolicy in store, instead returns an
@@ -234,24 +133,24 @@ func (n *NetworkPolicyController) processAntreaNetworkPolicy(np *secv1alpha1.Net
 	// Compute NetworkPolicyRule for Egress Rule.
 	for idx, ingressRule := range np.Spec.Ingress {
 		// Set default action to ALLOW to allow traffic.
-		action := ingressRule.Action.ToUpper()
+		services, namedPortExists := toAntreaServicesForCRD(ingressRule.Ports)
 		rules = append(rules, networking.NetworkPolicyRule{
 			Direction: networking.DirectionIn,
-			From:      *n.toAntreaPeerForCRD(ingressRule.From, np, networking.DirectionIn),
-			Services:  toAntreaServicesForCRD(ingressRule.Ports),
-			Action:    &action,
+			From:      *n.toAntreaPeerForCRD(ingressRule.From, np, networking.DirectionIn, namedPortExists),
+			Services:  services,
+			Action:    ingressRule.Action,
 			Priority:  int32(idx),
 		})
 	}
 	// Compute NetworkPolicyRule for Egress Rule.
 	for idx, egressRule := range np.Spec.Egress {
 		// Set default action to ALLOW to allow traffic.
-		action := egressRule.Action.ToUpper()
+		services, namedPortExists := toAntreaServicesForCRD(egressRule.Ports)
 		rules = append(rules, networking.NetworkPolicyRule{
 			Direction: networking.DirectionOut,
-			To:        *n.toAntreaPeerForCRD(egressRule.To, np, networking.DirectionOut),
-			Services:  toAntreaServicesForCRD(egressRule.Ports),
-			Action:    &action,
+			To:        *n.toAntreaPeerForCRD(egressRule.To, np, networking.DirectionOut, namedPortExists),
+			Services:  services,
+			Action:    egressRule.Action,
 			Priority:  int32(idx),
 		})
 	}
