+++ pkg/controller/networkpolicy/networkpolicy_crd_utils.go
@@ -1,0 +1,130 @@
+// Copyright 2020 Antrea Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package networkpolicy
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	"k8s.io/klog"
+
+	"github.com/vmware-tanzu/antrea/pkg/apis/networking"
+	secv1alpha1 "github.com/vmware-tanzu/antrea/pkg/apis/security/v1alpha1"
+	antreatypes "github.com/vmware-tanzu/antrea/pkg/controller/types"
+)
+
+var (
+	// matchAllPodsPeerCrd is a secv1alpha1.NetworkPolicyPeer matching all
+	// Pods from all Namespaces.
+	matchAllPodsPeerCrd = secv1alpha1.NetworkPolicyPeer{
+		NamespaceSelector: &metav1.LabelSelector{},
+	}
+)
+
+// toAntreaServicesForCRD converts a slice of secv1alpha1.NetworkPolicyPort
+// objects to a slice of Antrea Service objects. A bool is returned along with
+// the Service objects to indicate whether any named port exists.
+func toAntreaServicesForCRD(npPorts []secv1alpha1.NetworkPolicyPort) ([]networking.Service, bool) {
+	var antreaServices []networking.Service
+	var namedPortExists bool
+	for _, npPort := range npPorts {
+		if npPort.Port != nil && npPort.Port.Type == intstr.String {
+			namedPortExists = true
+		}
+		antreaService := networking.Service{
+			Protocol: toAntreaProtocol(npPort.Protocol),
+			Port:     npPort.Port,
+		}
+		antreaServices = append(antreaServices, antreaService)
+	}
+	return antreaServices, namedPortExists
+}
+
+// toAntreaIPBlockForCRD converts a secv1alpha1.IPBlock to an Antrea IPBlock.
+func toAntreaIPBlockForCRD(ipBlock *secv1alpha1.IPBlock) (*networking.IPBlock, error) {
+	// Convert the allowed IPBlock to networkpolicy.IPNet.
+	ipNet, err := cidrStrToIPNet(ipBlock.CIDR)
+	if err != nil {
+		return nil, err
+	}
+	antreaIPBlock := &networking.IPBlock{
+		CIDR: *ipNet,
+		// secv1alpha.IPBlock does not have the Except slices.
+		Except: []networking.IPNet{},
+	}
+	return antreaIPBlock, nil
+}
+
+func (n *NetworkPolicyController) toAntreaPeerForCRD(peers []secv1alpha1.NetworkPolicyPeer,
+	np secv1alpha1.NetworkPolicyType, dir networking.Direction, namedPortExists bool) *networking.NetworkPolicyPeer {
+	var addressGroups []string
+	// Empty NetworkPolicyPeer is supposed to match all addresses.
+	// It's treated as an IPBlock "0.0.0.0/0".
+	if len(peers) == 0 {
+		// For an egress Peer that specifies any named ports, it creates or
+		// reuses the AddressGroup matching all Pods in all Namespaces and
+		// appends the AddressGroup UID to the returned Peer such that it can be
+		// used to resolve the named ports.
+		// For other cases it uses the IPBlock "0.0.0.0/0" to avoid the overhead
+		// of handling member updates of the AddressGroup.
+		if dir == networking.DirectionIn || !namedPortExists {
+			return &matchAllPeer
+		}
+		allPodsGroupUID := n.createAddressGroupForCRD(matchAllPodsPeerCrd, np)
+		podsPeer := matchAllPeer
+		podsPeer.AddressGroups = append(addressGroups, allPodsGroupUID)
+		return &podsPeer
+	}
+	var ipBlocks []networking.IPBlock
+	for _, peer := range peers {
+		// A secv1alpha1.NetworkPolicyPeer will either have an IPBlock or a
+		// podSelector and/or namespaceSelector set.
+		if peer.IPBlock != nil {
+			ipBlock, err := toAntreaIPBlockForCRD(peer.IPBlock)
+			if err != nil {
+				klog.Errorf("Failure processing Antrea NetworkPolicy %s/%s IPBlock %v: %v", np.GetNamespace(), np.GetName(), peer.IPBlock, err)
+				continue
+			}
+			ipBlocks = append(ipBlocks, *ipBlock)
+		} else {
+			normalizedUID := n.createAddressGroupForCRD(peer, np)
+			addressGroups = append(addressGroups, normalizedUID)
+		}
+	}
+	return &networking.NetworkPolicyPeer{AddressGroups: addressGroups, IPBlocks: ipBlocks}
+}
+
+// createAddressGroupForCRD creates an AddressGroup object corresponding to a
+// secv1alpha1.NetworkPolicyPeer object in Antrea NetworkPolicyRule. This
+// function simply creates the object without actually populating the
+// PodAddresses as the affected Pods are calculated during sync process.
+func (n *NetworkPolicyController) createAddressGroupForCRD(peer secv1alpha1.NetworkPolicyPeer, np secv1alpha1.NetworkPolicyType) string {
+	groupSelector := toGroupSelector(np.GetNamespace(), peer.PodSelector, peer.NamespaceSelector)
+	normalizedUID := getNormalizedUID(groupSelector.NormalizedName)
+	// Get or create an AddressGroup for the generated UID.
+	_, found, _ := n.addressGroupStore.Get(normalizedUID)
+	if found {
+		return normalizedUID
+	}
+	// Create an AddressGroup object per Peer object.
+	addressGroup := &antreatypes.AddressGroup{
+		UID:      types.UID(normalizedUID),
+		Name:     normalizedUID,
+		Selector: *groupSelector,
+	}
+	klog.V(2).Infof("Creating new AddressGroup %s with selector (%s)", addressGroup.Name, addressGroup.Selector.NormalizedName)
+	n.addressGroupStore.Create(addressGroup)
+	return normalizedUID
+}
